Update README.md
The project aims to conduct a sentiment analysis of the ongoing Ukraine-Russia war. By analyzing social media posts, news articles, and public opinions, the project seeks to understand and quantify the sentiment surrounding the conflict. The insights gained from this analysis will contribute to a better understanding the emotional perspectives and attitudes toward the war from different stakeholders and the general public.